requirements.txt
# Core ML and Data Processing
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
joblib==1.3.2

# Real-time streaming and messaging
kafka-python==2.0.2
redis==4.6.0
websockets==11.0.2

# Web framework and API
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
requests==2.31.0

# Database
sqlalchemy==2.0.20

# Monitoring and logging
mlflow==2.6.0
prometheus-client==0.17.1

# Frontend
streamlit==1.26.0
plotly==5.15.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
asyncio-mqtt==0.13.0

# Development
pytest==7.4.2
black==23.7.0
flake8==6.0.0
